% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{discard_blanks}
\alias{discard_blanks}
\title{Discard blanks}
\usage{
discard_blanks(txt)
}
\arguments{
\item{txt}{A character vector.}
}
\value{
A character vector.
}
\description{
Discard blanks
}
\examples{
discard_blanks(c("hello", "", "world", "", "!", ""))
}
