% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{extract_between}
\alias{extract_between}
\title{Extract text between patterns}
\usage{
extract_between(txt, before = NULL, after = NULL, include = TRUE)
}
\arguments{
\item{txt}{A character vector}

\item{before, after}{Patterns which surround the text to extract}

\item{include}{Should element containing `before` and `after` be included?}
}
\value{
The extracted text
}
\description{
Extract text between patterns
}
\examples{
extract_between(month.abb, "Fe", "Oc")
extract_between(month.abb, "Fe", "Oc", include = FALSE)
}
