% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{num4}
\alias{num4}
\title{Convert to numeric and keep 4 digits.}
\usage{
num4(x)
}
\arguments{
\item{x}{A numeric-like vector.}
}
\value{
A numeric vector.
}
\description{
Convert to numeric and keep 4 digits.
}
\examples{
num4(c("1", "3.141593"))
}
