% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{raw_receipt}
\alias{raw_receipt}
\title{Read receipt from PDF}
\usage{
raw_receipt(pdf)
}
\arguments{
\item{pdf}{File path or raw vector with pdf data.}
}
\value{
A raw receipt as a character vector.
}
\description{
Read receipt from PDF
}
