% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_receipt.R
\name{read_receipt}
\alias{read_receipt}
\title{Read a receipt from PDF}
\usage{
read_receipt(pdf)
}
\arguments{
\item{pdf}{File path or raw vector with pdf data.}
}
\value{
A tibble.
}
\description{
Read a receipt from PDF
}
\details{
`read_receipt` is vectorised over `pdf`.
}
