% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts.R
\name{get_adress}
\alias{get_adress}
\alias{get_date}
\alias{get_total}
\alias{get_discount}
\alias{get_topay}
\alias{get_discounts}
\title{Extract specific fields}
\usage{
get_adress(raw)

get_date(raw)

get_total(raw)

get_discount(raw)

get_topay(raw)

get_discounts(raw)
}
\arguments{
\item{raw}{A raw receipt.}
}
\value{
A vector or a tibble containing the desired information.
}
\description{
Extract specific fields
}
